+++ test/spec/sbt_version_spec.rb
@@ -4,11 +4,7 @@ require_relative 'spec_helper'
 
 describe 'Sbt version warnings' do
   it 'shows warning for sbt 0.x versions' do
-    new_default_hatchet_runner('sbt-minimal-scala-sample').tap do |app|
-      app.before_deploy do
-        File.write('project/build.properties', "sbt.version=0.13.18\n")
-      end
-
+    new_default_hatchet_runner('sbt-0.13.18-minimal-with-native-packager').tap do |app|
       app.deploy do
         expect(clean_output(app.output)).to include(<<~OUTPUT)
           remote:  !     Warning: Unsupported sbt version detected.
@@ -18,7 +14,7 @@ describe 'Sbt version warnings' do
           remote:  !     Support for sbt 0.x was ended by the upstream sbt project on November 30, 2018.
           remote:  !     
           remote:  !     Please upgrade to sbt 1.x for active support.
-          remote:  !
+          remote:  !     
           remote:  !     The buildpack will attempt to build your application, but compatibility
           remote:  !     is not guaranteed and may break at any time.
           remote:  !     
@@ -36,7 +32,7 @@ describe 'Sbt version warnings' do
   end
 
   it 'does not show warning for sbt 1.x versions' do
-    new_default_hatchet_runner('sbt-one-example').tap do |app|
+    new_default_hatchet_runner('sbt-1.11.7-minimal-with-native-packager').tap do |app|
       app.deploy do
         expect(clean_output(app.output)).not_to include('Warning: Unsupported sbt version detected.')
         expect(app.output).to include('[success]')
@@ -44,23 +40,50 @@ describe 'Sbt version warnings' do
     end
   end
 
-  it 'shows warning for sbt 2.x versions' do
-    new_default_hatchet_runner('sbt-minimal-scala-sample').tap do |app|
+  it 'shows error for sbt 2.x versions' do
+    new_default_hatchet_runner('sbt-2.0.0-RC6-minimal-with-native-packager', allow_failure: true).tap do |app|
+      app.deploy do
+        expect(app).not_to be_deployed
+        expect(clean_output(app.output)).to include(<<~OUTPUT)
+          remote:  !     Error: Unsupported sbt version detected.
+          remote:  !     
+          remote:  !     This buildpack does not currently support sbt 2.x. You are using sbt 2.0.0-RC6.
+          remote:  !     
+          remote:  !     Support for sbt 2.x will be added in a future buildpack release. In the
+          remote:  !     meantime, please use the latest stable sbt 1.x version for your deployments.
+          remote:  !     
+          remote:  !     To continue, update project/build.properties to use sbt 1.x.
+          remote:  !     
+          remote:  !     For more information:
+          remote:  !     - Latest sbt 1.x releases: https://github.com/sbt/sbt/releases
+          remote:  !     - sbt 2.x changes: https://www.scala-sbt.org/2.x/docs/en/changes/sbt-2.0-change-summary.html
+        OUTPUT
+      end
+    end
+  end
+
+  it 'shows error when sbt version cannot be determined' do
+    new_default_hatchet_runner('sbt-1.11.7-minimal-with-native-packager', allow_failure: true).tap do |app|
       app.before_deploy do
-        File.write('project/build.properties', "sbt.version=2.0.0-M1\n")
+        # Remove contents of project/build.properties
+        File.write('project/build.properties', '')
       end
 
       app.deploy do
+        expect(app).not_to be_deployed
         expect(clean_output(app.output)).to include(<<~OUTPUT)
-          remote:  !     Warning: Unsupported sbt version detected.
+          remote:  !     Error: sbt version cannot be determined.
           remote:  !     
-          remote:  !     This buildpack does not support sbt 2.x yet. You are using sbt 2.0.0-M1.
-          remote:  !     Note that sbt 2.x is still in development.
+          remote:  !     As part of your build definition you must specify the version of sbt that
+          remote:  !     your build uses. This ensures consistent results across different environments.
           remote:  !     
-          remote:  !     Please use sbt 1.x for stable, production deployments.
+          remote:  !     To fix this issue, create a file named project/build.properties that specifies
+          remote:  !     the sbt version as follows:
           remote:  !     
-          remote:  !     The buildpack will attempt to build your application, but compatibility
-          remote:  !     is not guaranteed and may break at any time.
+          remote:  !     	sbt.version=1.11.7
+          remote:  !     
+          remote:  !     For more information, see:
+          remote:  !     https://www.scala-sbt.org/1.x/docs/Basic-Def.html#Specifying+the+sbt+version
         OUTPUT
       end
     end